[ REFACTOR ] 기존 popver방식을 일반 버튼 나열로 변경 (#87)
Why
- 기존의 popover방식은 번거로운 방식이라 변경해야 한다.

What
- 바로 이모지를 날릴 수 있도록 버튼에 나열하는 방식으로 변경.
- 기존의 popover 삭제.
