--- web/src/components/channel/Slide/SlideInfo/FlyingEmojiPopover/index.js
@@ -1,1 +1,0 @@
-export { default } from './FlyingEmojiPopover';
