--- web/src/components/channel/Slide/SlideInfo/FlyingEmojiPopover/style.jsx
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-import { px } from '@/styles/themeUtil';
-
-export default {
-  EmojiButton: styled.span`
-    font-size: ${px(30)};
-    margin: ${px(5)};
-    width: 50%;
-    cursor: pointer;
-  `,
-};
